Move soroban metrics into a new file
